fix: fix -Wextra warnings in ccl/, reduce_to_root/, conv2d/ (r26 batch)

+++ ttnn/cpp/ttnn/operations/ccl/ccl_op_fusion.cpp
@@ -420,7 +420,7 @@ void MatmulFusedOpSignaler::init_llama_rs_cores_mm(
     // pick the privileged core, record the number of matmul cores
     TT_FATAL(initialized_llama_reduce_scatter_part1, "reduce scatter half needs to be initialized first");
     auto cores = corerange_to_cores(matmul_cores);
-    TT_FATAL(cores.size() > privilaged_index, "Privileged index is out of range of the matmul cores");
+    TT_FATAL(cores.size() > static_cast<size_t>(privilaged_index), "Privileged index is out of range of the matmul cores");
     this->privilaged_core = cores.at(privilaged_index);
     this->privilaged_core_physical = device->worker_core_from_logical_core(this->privilaged_core);
     this->matmul_privilaged_semaphore = tt::tt_metal::CreateSemaphore(program, privilaged_core, 0);
